+++ packages/core/src/GlobalForm.tsx
@@ -1,9 +1,12 @@
 import React from 'react'
 import { Form } from '@i-vresse/wb-form'
-import { useSetActiveSubmitButton, useCatalog, useGlobalFormData } from './store'
+
+import { FormProps } from './FormProps'
+import { useCatalog, useGlobalFormData, useSetActiveSubmitButton } from './store'
+
 import '@i-vresse/wb-form/dist/index.css'
 
-export const GlobalForm = (): JSX.Element => {
+export const GlobalForm = ({ fields, widgets }: FormProps): JSX.Element => {
   const { global: globalSchemas } = useCatalog()
   const [formData, setFormData] = useGlobalFormData()
   const submitFormRefSetter = useSetActiveSubmitButton()
@@ -14,6 +17,8 @@ export const GlobalForm = (): JSX.Element => {
       uiSchema={uiSchema}
       formData={formData}
       onSubmit={({ formData }) => setFormData(formData)}
+      fields={fields}
+      widgets={widgets}
     >
       <button ref={submitFormRefSetter} style={{ display: 'none' }} />
     </Form>