+++ packages/core/src/NodePanel.tsx
@@ -1,19 +1,35 @@
 import React from 'react'
+
+import { FormProps } from './FormProps'
 import { GlobalForm } from './GlobalForm'
 import { NodeForm } from './NodeForm'
 import { useSelectNodeIndex, useWorkflow } from './store'
 
-export const NodePanel = (): JSX.Element => {
+/**
+ * Panel which renders the form for the selected node or the global parameters.
+ *
+ */
+export const NodePanel = ({ fields, widgets }: FormProps): JSX.Element => {
   const selectedNodeIndex = useSelectNodeIndex()
   const { editingGlobal } = useWorkflow()
   let form = <div>No node or global parameters selected for configuration.</div>
   let legend = 'Node'
   if (editingGlobal) {
-    form = <GlobalForm />
+    form = (
+      <GlobalForm
+        fields={fields}
+        widgets={widgets}
+      />
+    )
     legend = 'Global parameters'
   }
   if (selectedNodeIndex !== -1) {
-    form = <NodeForm />
+    form = (
+      <NodeForm
+        fields={fields}
+        widgets={widgets}
+      />
+    )
   }
   return (
     <fieldset>