+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## @i-vresse/wb-core 1.1.2 - 2022-08-23
+
+### Added
+
+* Allow custom fields and widgets in form. ([#116](https://github.com/i-VRESSE/workflow-builder/pull/116))
+
 ## @i-vresse/wb-form 1.1.2 - 2022-08-12
 
 ### Added